3.5 fix / Missing sender
Fixed code to use 3.5-rc6 Sharp libraries, as changes made to SetMouseMode() has been made, to use MouseMode, and MouseModeEnum.

Fixed bug in ConnectSignalHandler() where the signal attribute has an empty string for sender, it is referencing this, and it was not properly setting the sender field to this when initiating the Connect() call.