Updated to Godot 4 Beta 10 And Fixes
Updated version to Godot 4 Beta 10.

Fixed issue in TryGetNode()'s method of fetching the node, as node.HasNode() may not work with % Scene Unique Names.  Instead, we use node.GetNodeOrNull() to get the node, and if it returns null, we continue through the list, otherwise we return the node.

Also updated the code to remove all null strings, instead of just the first one found, and if name1 is an empty string  in AssignPathToMember(), it will use the empty string, instead of trying to place the Scene Unique Name % sign.