Added New Features
Added function to check if the variables that have ResolveNodeAttribute, NodePathAttribute, ResourceAttribute and SingletonAttribute, have non-null values set, if so, return true, else return false.
Added SceneNodeAttribute to store the resource path for the scene associated with class.
Created SceneNode<T>.FromScene() class, to allow easy creation of nodes that have a Scene associated with it.  Requires the path to the Scene to be associated with the Class using SceneNodeAttribute.

Updated Readme Documentation for functionality.